prompt users before deleting items from database

+++ scripts/extensions/broadcasting/src/rundowns/manage-rundown-items.tsx
@@ -114,7 +114,9 @@ export class ManageRundownItems<T extends IRundownItemBase | IRundownItem> exten
                         const deleteAction: IMenuItem = {
                             label: gettext('Delete'),
                             onSelect: () => {
-                                this.props.onDelete(item);
+                                superdesk.ui.confirm(gettext('Are you sure you want to delete it?')).then(() => {
+                                    this.props.onDelete(item);
+                                });
                             },
                         };
